Get rid of "types/interfaces" package.
